requirements.txt
google-genai
google-cloud-aiplatform
google-cloud-firestore
numpy
fastapi
uvicorn[standard]
websockets
python-dotenv
python-multipart
